Navbar Component Added
I created the components for all of the pages for our web app. For now all of the links say the name of the link they are in. However, once we develop each site we can add onto it. I added the bootstrap link but I don't think its necessary so I might remove it later to see if it still works. I was running into a 404 error when my href's were anything but " / " because I thought you would have to change it as you click each link but you do not. I have not finished the css and the look for it but that should be done by next week at the latest.