Manually Added David and Oscar's Changes to Development
Added David's work on the Calendar page and Oscar's
work on sending data from the front-end to Firestore
manually due to merge conflicts on their local machines.
We are trying to fix a conflict issue, and ultimately
update Development to be in a state where everyone
can branch from and have the most recent version of
our app.